fix build&test
